typo, change useDuration hook to return time = 0 when the order has ended, moved action function from oders.tsx -> order._index.tsx
