+++ WebPage/app/components/common/OrderPage/OrdersTableHeader.tsx
@@ -50,7 +50,7 @@ export default function OrderTableHeader({
       <tr>
         <OrderTableHeaderCell
           column="orderedBy"
-          title="Užsakymas sukurė"
+          title="Užsakymą sukurė"
           handleSort={handleSort}
           sortOrder={sortOrder}
           sortColumn={sortColumn}
