+++ WebPage/app/hooks/useDuration.ts
@@ -2,7 +2,12 @@ import { useEffect, useState } from "react";
 
 export default function useDuration(completionDate: Date) {
   const calculateTimeRemaining = () => {
-    const endInMs = completionDate.getTime() - new Date().getTime();
+    const currentDate = new Date();
+
+    if (completionDate <= currentDate)
+      return { seconds: 0, minutes: 0, hours: 0, days: 0, endInMs: 0 };
+
+    const endInMs = completionDate.getTime() - currentDate.getTime();
 
     let seconds = Math.trunc(Math.abs(endInMs) / 1000);
